requirements.txt
git+https://github.com/basxsoftwareassociation/excel2django
git+https://github.com/basxsoftwareassociation/htmlgenerator
git+https://github.com/basxsoftwareassociation/bread
django-phonenumber-field
phonenumbers
# using a forked version because the public release has a backward incompatability with django 2
git+https://github.com/basxltd/django-language-field
django-formtools
